major update
this commit adds a lot of unrelated features. it adds proper support for
scope based launching of apps -- also the ones that don't have proper
dbus activation. it changes the description type to be a widget
generator. it also enhances the performance of the search quite a bit.
the hashset has been refactored a lot, and duplicate detection is now
really fast.